fix: apply required fixes from code review

+++ firstassignment/src/controllers/expense.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, Param, Post, Put } from '@nestjs/common';
+import { Body, Controller, Delete, Get, NotFoundException, Param, Post, Put } from '@nestjs/common';
 import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
 import { ExpensesService } from '../services/expense.service';
 import { CreateExpenseDto } from '../Dto/create-expense-dto';
@@ -11,24 +11,28 @@ export class ExpenseController {
         private readonly expenseService: ExpensesService, 
     ) {}
 
-    @Get('expense/:id')
+    @Get('expenseById/:id')
     @ApiOperation({description: "Zwraca wydatek"})
-    @ApiResponse({ status: 201, description: "Sukces!"})
+    @ApiResponse({ status: 200, description: "Sukces!"})
     async getWholeExpense(@Param('id') id: number){
-        return this.expenseService.expense(id);
+        return this.expenseService.expenseById(id);
     }
-
-    @Get('expenses')
+ 
+    @Get('allExpenses')
     @ApiOperation({description: "Zwraca wszystkie wydatki"})
-    @ApiResponse({ status: 201, description: "Sukces!"})
+    @ApiResponse({ status: 200, description: "Sukces!"})
     async getWholeExpenses(){
-        return this.expenseService.expenses();
+        return this.expenseService.allExpenses();
     }
 
     @Delete('deleteExpense/:id')
     @ApiOperation({description: "Usuwa wybrany wydatek"})
-    @ApiResponse({ status: 201, description: "Sukces!"})
+    @ApiResponse({ status: 200, description: "Sukces!"})
     async deleteExpense(@Param('id') id: string){
+        const expense = await this.expenseService.expenseById(Number.parseInt(id));
+        if(expense === null){
+            throw new NotFoundException(`Wydatek z ID ${id} nie istnieje`);
+        }
         return this.expenseService.deleteExpense(Number.parseInt(id));
     }
 
@@ -40,9 +44,13 @@ export class ExpenseController {
     }
 
     @Put('updateExpense/:id')
-    @ApiOperation({description: "Dodaje nowy wydatek"})
-    @ApiResponse({ status: 201, description: "Sukces!"})
+    @ApiOperation({description: "Aktualizuje wydatek"})
+    @ApiResponse({ status: 200, description: "Sukces!"})
     async updateExpense(@Param('id') id: string, @Body() newData: CreateExpenseDto){
+        const expense = await this.expenseService.expenseById(Number.parseInt(id));
+        if(expense === null){
+            throw new NotFoundException(`Wydatek z ID ${id} nie istnieje`);
+        }
         const parameters = {id: Number.parseInt(id), newData}
         return this.expenseService.updateExpense(parameters);
     }
