+++ firstassignment/src/services/participant.service.ts
@@ -1,25 +1,27 @@
-import { Injectable } from "@nestjs/common";
+import { Injectable, NotFoundException } from "@nestjs/common";
 import { PrismaService } from "./prisma.service";
 import { Participant } from "../../generated/prisma";
 import { CreateParticipantDto } from "src/Dto/create-participant-dto";
 import { TripsService } from "./trip.service";
+import { UpdateParticipantDto } from "src/Dto/update-participant-dto";
 
 @Injectable()
 export class ParticipantsService{
     constructor(private prisma: PrismaService, private tripService: TripsService){}
 
-    async participant(id: number): Promise<Participant>{
-        return this.prisma.participant.findFirstOrThrow({where: {id},include: {expenses: true}});
+    async participantById(id: number){
+        return this.prisma.participant.findUnique({where: {id},include: {expenses: true}});
     }
 
-    async participants(): Promise<Participant[]>{
+    async allParticipants(): Promise<Participant[]>{
         return this.prisma.participant.findMany();
     }
 
-    async createParticipant(data: CreateParticipantDto): Promise<void> {
-        await this.tripService.trip(data.tripId)
-        .catch((error: unknown) => {console.error(error)})
-        .finally(async () => {
+    async createParticipant(data: CreateParticipantDto){
+        const trip = await this.tripService.tripById(data.tripId);
+        if(trip === null){
+            throw new NotFoundException(`Wyacieczka z ID ${data.tripId.toString()} nie istnieje`);
+        }else{
             return this.prisma.participant.create({
                 data:{
                     imie: data.imie,
@@ -31,15 +33,26 @@ export class ParticipantsService{
                     updated_at: new Date(),
                 }
             })
-        });
+        }
     }
     
     async updateParticipant(parameters:{
             id: number;
-            newData: CreateParticipantDto;
-    }): Promise<Participant>{
+            newData: UpdateParticipantDto;
+    }){
         const { id, newData } = parameters;
-        return this.prisma.participant.update({
+        const participant = await this.participantById(id);
+        if(newData.tripId !== undefined){
+            const trip = await this.tripService.tripById(newData.tripId);
+            if(trip === null){
+                throw new NotFoundException(`Wycieczka z ID ${newData.tripId.toString()} nie istnieje`)
+            }
+        }
+        if(participant === null){
+            throw new NotFoundException(`Uczestnik z ID ${id.toString()} nie istnieje`);
+        }
+        else{
+            return this.prisma.participant.update({
             data: {
                 imie: newData.imie,
                 nazwisko: newData.nazwisko,
@@ -50,9 +63,16 @@ export class ParticipantsService{
             },
             where: {id},
         })
+        }
     }
 
-    async deleteParticipant(id: number): Promise<Participant>{
-        return this.prisma.participant.delete({where: {id}, include: {expenses: true}})
+    async deleteParticipant(id: number){
+        const participant = await this.participantById(id);
+        if(participant === null){
+            throw new NotFoundException(`Uczestnik z ID ${id.toString()} nie istnieje`);
+        }
+        else{
+            return this.prisma.participant.delete({where: {id}, include: {expenses: true}})
+        }
     }
 }
